+++ tutorials/n5geh_cluster/t3_deploy_controller/e5_iot_thermal_zone_control_solution.py
