+++ tutorials/n5geh_cluster/t3_deploy_controller/zone_temperature_sensor.png
